Code cleanup Removed capital S where found changed pin config option to mcp23sxx_id: updated test1.yaml
