+++ esphome/components/mcp23s08/mcp23s08.h
@@ -45,14 +45,13 @@ class MCP23S08 : public Component,
 
   float get_setup_priority() const override;
 
-  // read a given register
-  bool read_reg(uint8_t reg, uint8_t *value);
-  // write a value to a given register
-  bool write_reg(uint8_t reg, uint8_t value);
-  // update registers with given pin value.
-  void update_reg(uint8_t pin, bool pin_value, uint8_t reg_a);
-
  protected:
+  // read a given register
+  bool read_reg_(uint8_t reg, uint8_t *value);
+  // write a value to a given register
+  bool write_reg_(uint8_t reg, uint8_t value);
+  // update registers with given pin value.
+  void update_reg_(uint8_t pin, bool pin_value, uint8_t reg_a);
   uint8_t device_opcode_ = 0x40;
   uint8_t olat_{0x00};
 };
